feat: Add socket.io dependency and implement leaveGroup function
This commit adds the "socket.io" dependency to the project's package.json file. It also implements the "leaveGroup" function in the "groupChatController.js" file. The "leaveGroup" function allows a user to leave a group conversation by removing their ID from the conversation's members array. If the user is the last member of the conversation, the conversation is deleted.

These changes enhance the chat functionality by enabling users to leave group conversations and keeping the dependencies up to date.